モードの実装
