+++ include/libwrap/LibWrapper.hpp
@@ -1,0 +1,10 @@
+#ifdef LIB_WRAPPER_HPP
+#define LIB_WRAPPER_HPP
+
+namespace wrap{
+/* 0 ncurses */
+#define UI_NCURSES 0
+#define INPUT_NCURSES 0
+}
+
+#endif
